Move `ConfigData` to separate file
This will help with improving its interface.
