+++ mk/libboat.deps
@@ -1,1 +1,1 @@
-src/configcontainer.cpp src/configparser.cpp src/colormanager.cpp src/keymap.cpp src/stflpp.cpp src/logger.cpp src/exception.cpp src/utils.cpp src/fslock.cpp src/matcher.cpp src/fmtstrformatter.cpp src/strprintf.cpp src/confighandlerexception.cpp src/matcherexception.cpp src/scopemeasure.cpp src/history.cpp src/ruststring.cpp
+src/configdata.cpp src/configcontainer.cpp src/configparser.cpp src/colormanager.cpp src/keymap.cpp src/stflpp.cpp src/logger.cpp src/exception.cpp src/utils.cpp src/fslock.cpp src/matcher.cpp src/fmtstrformatter.cpp src/strprintf.cpp src/confighandlerexception.cpp src/matcherexception.cpp src/scopemeasure.cpp src/history.cpp src/ruststring.cpp
