+++ mk/mk.deps
@@ -75,9 +75,11 @@ src/colormanager.o: src/colormanager.cpp include/colormanager.h \
  include/urlviewformaction.h include/utils.h include/logger.h
 src/configcontainer.o: src/configcontainer.cpp include/configcontainer.h \
  include/configparser.h include/configactionhandler.h config.h \
- include/configparser.h include/confighandlerexception.h include/logger.h \
- include/strprintf.h include/strprintf.h include/utils.h \
- 3rd-party/optional.hpp include/configcontainer.h include/logger.h
+ include/configparser.h include/configdata.h \
+ include/confighandlerexception.h include/logger.h include/strprintf.h \
+ include/strprintf.h include/utils.h 3rd-party/optional.hpp \
+ include/configcontainer.h include/logger.h
+src/configdata.o: src/configdata.cpp include/configdata.h
 src/confighandlerexception.o: src/confighandlerexception.cpp \
  include/confighandlerexception.h include/configparser.h \
  include/configactionhandler.h config.h
@@ -648,7 +650,7 @@ test/colormanager.o: test/colormanager.cpp include/colormanager.h \
  include/confighandlerexception.h
 test/configcontainer.o: test/configcontainer.cpp \
  include/configcontainer.h include/configparser.h \
- include/configactionhandler.h 3rd-party/catch.hpp \
+ include/configactionhandler.h 3rd-party/catch.hpp include/configdata.h \
  include/confighandlerexception.h include/configparser.h include/keymap.h
 test/configparser.o: test/configparser.cpp include/configparser.h \
  include/configactionhandler.h 3rd-party/catch.hpp include/keymap.h \
