+++ src/configcontainer.cpp
@@ -9,6 +9,7 @@
 
 #include "config.h"
 #include "configparser.h"
+#include "configdata.h"
 #include "confighandlerexception.h"
 #include "logger.h"
 #include "strprintf.h"
